Move `util` to `util.misc`
Also move the `oneline` function to its own module.

This is the first step to breaking out parts of `util.misc` into more
specialized modules.
