+++ bionic/cache_api.py
@@ -6,8 +6,8 @@ with Bionic's cache.
 from functools import total_ordering
 
 from .descriptors.parsing import entity_dnode_from_descriptor
+from .oneline import oneline
 from .persistence import path_from_url, is_file_url
-from .util import oneline
 
 
 class Cache:
