+++ bionic/descriptors/parsing.py
@@ -28,7 +28,7 @@ import attr
 
 from .ast import EntityNode, TupleNode
 from ..exception import MalformedDescriptorError
-from ..util import oneline
+from ..oneline import oneline
 
 
 def dnode_from_descriptor(descriptor):
